[REFACTOR]: add_page_0 파일 구조 변경
